requirements.txt
langgraph==0.2.34
openai==1.51.2
python-telegram-bot==21.5
python-dotenv==1.0.1
asyncio
typing-extensions
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytest==8.3.4
pytest-asyncio==0.23.8
pytest-mock==3.14.0
freezegun==1.5.1